cleanup package info and use correct latest versions
