+++ .github/workflows/update-supported-versions-doc.yml
@@ -1,10 +1,12 @@
-name: Supported Versions
+name: Update supported integration versions doc
 on:
-    workflow_run:
-        workflows: ["Smoke Tests"]
-        branches: [main]
-        types: 
-            - completed
+  workflow_run:
+    # Every time we run smoke tests, we check what version of each library was used to run the tests and this will
+    # be used as the latest supported version in our docs.
+    workflows: [ "Smoke Tests" ]
+    branches: [ main ]
+    types:
+      - completed
 
 concurrency:
   # Automatically cancel previous runs if a new one is triggered to conserve resources.
@@ -19,19 +21,13 @@ jobs:
       contents: write
       pull-requests: write
     env:
-        GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
+      GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
     if: ${{ github.event.workflow_run.conclusion == 'success' }}
     steps:
       - uses: actions/checkout@v4
-
       - run: echo 'Smoke Tests workflow passed'
-
-      - run: go run .github/workflows/apps/output_supported_versions.go 
-
-      - run: go mod tidy
-      
+      - run: go run .github/workflows/apps/gen_supported_versions_doc.go
       - run: git diff
-
       # note: This will only run when there *are* changes to integration versions
       - name: Create Pull Request
         id: pr
@@ -39,8 +35,8 @@ jobs:
         with:
           token: ${{ secrets.GITHUB_TOKEN }}
           branch: "update-latest-supported-versions"
-          commit-message: "Update supported versions"
+          commit-message: "docs: update latest supported integration versions doc"
           base: main
-          title: "chore: update latest supported"
+          title: "docs: update latest supported integration versions doc"
           labels: changelog/no-changelog
           body: "Test"
