+++ contrib/supported_versions.md
@@ -1,0 +1,54 @@
+| Dependency |    Repository  | Minimum Version | Maximum Version | Auto-Instrumented |
+|------------|-----------------|-----------------|-----------------|-----------------|
+| cloud.google.com/go/pubsub | cloud.google.com/go/pubsub | v1.36.1 | v1.45.3 | true |
+| database/sql | database/sql | - | - | true |
+| github.com/99designs/gqlgen | github.com/99designs/gqlgen | v0.17.36 | v0.17.63 | true |
+| github.com/IBM/sarama | github.com/IBM/sarama | v1.40.0 | v1.45.0 | true |
+| github.com/Shopify/sarama | github.com/Shopify/sarama | v1.38.1 | v1.45.0 | true |
+| github.com/aws/aws-sdk-go | github.com/aws/aws-sdk-go | v1.44.327 | v1.55.5 | true |
+| github.com/aws/aws-sdk-go-v2 | github.com/aws/aws-sdk-go-v2 | v1.20.3 | v1.32.8 | true |
+| github.com/bradfitz/gomemcache | github.com/bradfitz/gomemcache | v0.0.0-20230611145640-acc696258285 | v0.0.0-20230905024940-24af94b03874 | false |
+| github.com/confluentinc/confluent-kafka-go | github.com/confluentinc/confluent-kafka-go | v1.9.2 | v1.9.2 | false |
+| github.com/confluentinc/confluent-kafka-go/v2 | github.com/confluentinc/confluent-kafka-go/v2 | v2.2.0 | v2.8.0 | false |
+| github.com/dimfeld/httptreemux/v5 | github.com/dimfeld/httptreemux/v5 | v5.5.0 | v5.5.0 | false |
+| github.com/elastic/go-elasticsearch/v6 | github.com/elastic/go-elasticsearch/v6 | v6.8.5 | v6.8.10 | true |
+| github.com/emicklei/go-restful/v3 | github.com/emicklei/go-restful/v3 | v3.11.0 | v3.12.1 | false |
+| github.com/envoyproxy/go-control-plane | github.com/envoyproxy/go-control-plane | v0.13.1 | v0.13.4 | false |
+| github.com/gin-gonic/gin | github.com/gin-gonic/gin | v1.9.1 | v1.10.0 | true |
+| github.com/globalsign/mgo | github.com/globalsign/mgo | v0.0.0-20181015135952-eeefdecb41b8 | v0.0.0-20181015135952-eeefdecb41b8 | false |
+| github.com/go-chi/chi | github.com/go-chi/chi | v1.5.4 | v1.5.5 | true |
+| github.com/go-chi/chi/v5 | github.com/go-chi/chi/v5 | v5.0.10 | v5.2.0 | true |
+| github.com/go-pg/pg/v10 | github.com/go-pg/pg/v10 | v10.11.1 | v10.14.0 | false |
+| github.com/go-redis/redis | github.com/go-redis/redis | v6.15.9+incompatible | v6.15.9+incompatible | false |
+| github.com/go-redis/redis/v7 | github.com/go-redis/redis/v7 | v7.4.1 | v7.4.1 | true |
+| github.com/go-redis/redis/v8 | github.com/go-redis/redis/v8 | v8.11.5 | v8.11.5 | true |
+| github.com/gocql/gocql | github.com/gocql/gocql | v1.6.0 | v1.7.0 | true |
+| github.com/gofiber/fiber/v2 | github.com/gofiber/fiber/v2 | v2.52.5 | v2.52.6 | true |
+| github.com/gomodule/redigo | github.com/gomodule/redigo | v1.8.9 | v1.9.2 | false |
+| github.com/gorilla/mux | github.com/gorilla/mux | v1.8.0 | v1.8.1 | true |
+| github.com/graph-gophers/graphql-go | github.com/graph-gophers/graphql-go | v1.5.0 | v1.5.0 | true |
+| github.com/graphql-go/graphql | github.com/graphql-go/graphql | v0.8.1 | v0.8.1 | true |
+| github.com/hashicorp/consul/api | github.com/hashicorp/consul/api | v1.24.0 | v1.31.0 | false |
+| github.com/hashicorp/vault/api | github.com/hashicorp/vault/api | v1.9.2 | v1.15.0 | true |
+| github.com/jackc/pgx/v5 | github.com/jackc/pgx/v5 | v5.6.0 | v5.7.2 | true |
+| github.com/jmoiron/sqlx | github.com/jmoiron/sqlx | v1.3.5 | v1.4.0 | false |
+| github.com/julienschmidt/httprouter | github.com/julienschmidt/httprouter | v1.3.0 | v1.3.0 | true |
+| github.com/labstack/echo/v4 | github.com/labstack/echo/v4 | v4.11.1 | v4.13.3 | true |
+| github.com/miekg/dns | github.com/miekg/dns | v1.1.55 | v1.1.62 | false |
+| github.com/redis/go-redis/v9 | github.com/redis/go-redis/v9 | v9.1.0 | v9.7.0 | true |
+| github.com/segmentio/kafka-go | github.com/segmentio/kafka-go | v0.4.42 | v0.4.47 | true |
+| github.com/sirupsen/logrus | github.com/sirupsen/logrus | v1.9.3 | v1.9.3 | true |
+| github.com/syndtr/goleveldb | github.com/syndtr/goleveldb | v1.0.1-0.20220721030215-126854af5e6d | v1.0.1-0.20220721030215-126854af5e6d | false |
+| github.com/tidwall/buntdb | github.com/tidwall/buntdb | v1.3.0 | v1.3.2 | false |
+| github.com/twitchtv/twirp | github.com/twitchtv/twirp | v8.1.3+incompatible | v8.1.3+incompatible | true |
+| github.com/uptrace/bun | github.com/uptrace/bun | v1.1.17 | v1.2.8 | false |
+| github.com/urfave/negroni | github.com/urfave/negroni | v1.0.0 | v1.0.0 | false |
+| github.com/valyala/fasthttp | github.com/valyala/fasthttp | v1.51.0 | v1.58.0 | false |
+| go.mongodb.org/mongo-driver | go.mongodb.org/mongo-driver | v1.12.1 | v1.17.2 | false |
+| google.golang.org/api | google.golang.org/api | v0.169.0 | v0.216.0 | false |
+| google.golang.org/grpc | google.golang.org/grpc | v1.69.0 | v1.69.4 | true |
+| gopkg.in/olivere/elastic.v5 | gopkg.in/olivere/elastic.v5 | v5.0.84 | v5.0.86 | false |
+| gorm.io/gorm | gorm.io/gorm | v1.25.3 | v1.25.12 | true |
+| k8s.io/client-go | k8s.io/client-go | v0.25.5 | v0.32.0 | true |
+| log/slog | log/slog | - | - | true |
+| net/http | net/http | - | - | true |
